Improved analytics logging with middlware and routes
