+++ app/controllers/tweets.js
@@ -6,30 +6,7 @@ const User = mongoose.model("User");
 const Analytics = mongoose.model("Analytics");
 const _ = require("underscore");
 
-function logAnalytics(req) {
-  const url = req.protocol + "://" + req.get("host") + req.originalUrl;
-  if (req.ip !== undefined) {
-    const crudeIpArray = req.ip.split(":");
-    const ipArrayLength = crudeIpArray.length;
-    // cleanup IP to remove unwanted characters
-    const cleanIp = crudeIpArray[ipArrayLength - 1];
-    if (req.get("host").split(":")[0] !== "localhost") {
-      const analytics = new Analytics({
-        ip: cleanIp,
-        user: req.user,
-        url: url
-      });
-      analytics.save(err => {
-        if (err) {
-          console.log(err);
-        }
-      });
-    }
-  }
-}
-
 exports.tweet = (req, res, next, id) => {
-  logAnalytics(req);
   Tweet.load(id, (err, tweet) => {
     if (err) {
       return next(err);
@@ -44,7 +21,6 @@ exports.tweet = (req, res, next, id) => {
 
 // ### Create a Tweet
 exports.create = (req, res) => {
-  logAnalytics(req);
   const tweet = new Tweet(req.body);
   tweet.user = req.user;
   tweet.uploadAndSave({}, err => {
@@ -58,7 +34,6 @@ exports.create = (req, res) => {
 
 // ### Update a tweet
 exports.update = (req, res) => {
-  logAnalytics(req);
   let tweet = req.tweet;
   tweet = _.extend(tweet, {"body": req.body.tweet});
   tweet.uploadAndSave({}, (err) => {
@@ -71,9 +46,7 @@ exports.update = (req, res) => {
 
 // ### Delete a tweet
 exports.destroy = (req, res) => {
-  logAnalytics(req);
   const tweet = req.tweet;
-
   tweet.remove(err => {
     if (err) {
       return res.render("pages/500");
@@ -83,7 +56,6 @@ exports.destroy = (req, res) => {
 };
 
 exports.index = (req, res) => {
-  logAnalytics(req);
   const page = (req.param("page") > 0 ? req.param("page") : 1) - 1;
   const perPage = 10;
   const options = {
